implemented basic 'other' option in radio/checkboxes
