+++ app/assets/styles/partials/components/_field.scss
@@ -3,20 +3,55 @@
 }
 
 .field__item {
-  padding: 0.5rem;
-  margin-bottom: 0.5rem;
+  margin: 0 0.5rem 0.5rem 0;
   position: relative;
-  display: flex;
-  align-items: flex-start;
+  width: 100%;
+  min-width: 20rem;
+  display: inline-block;
+  @include mq(s) {
+    width: auto;
+  }
 }
 
-.field--multi {
+.field:not(.field--cols) {
+  .field__item:last-of-type {
+    margin-bottom: 0;
+  }
+}
+
+.field__other {
+  display: none;
+  clear: both;
+  padding: 0 0.5rem 0.5rem;
+}
+
+input:checked ~ .field__other {
+  display: block;
+  background: $color-lighter-grey;
+}
+
+.field--radio,
+.field--checkbox {
+  .field__item {
+    background: white;
+    border: 1px solid $color-borders;
+    border-radius: 0.2rem;
+    cursor: pointer;
+    &:hover {
+      border-color: $color-dark-grey;
+    }
+  }
+
   &.field--cols {
     @include mq(s) {
       display: flex;
       flex-wrap: wrap;
       .field__item {
-        width: 50%;
+        width: calc(50% - 1rem);
+        min-width: 0;
+      }
+      .label {
+        width: 100%;
       }
       &:last-of-type {
         margin-bottom: 0;
