+++ app/assets/styles/partials/components/_input.scss
@@ -45,16 +45,16 @@ $input-width: 15rem;
 .input--checkbox {
   width: 20px;
   height: 20px;
-  margin-top: 0.2rem;
   appearance: none;
   border: 2px solid #9b9b9b;
   padding: 0.5rem;
   font-size: 1rem;
   background: #eee;
   box-shadow: inset 0 0 0 3px white;
-  transition: background 300ms;
   cursor: pointer;
-  position: relative;
+  position: absolute;
+  top: 0.7rem;
+  left: 0.5rem;
   &:focus {
     border: 2px solid tint($color-blue, 50%);
     box-shadow: inset 0 0 0 3px white;
