+++ app/assets/styles/partials/components/_label.scss
@@ -5,52 +5,19 @@
   color: inherit;
 }
 
-.label__inner {
-  position: relative;
-  z-index: 3;
+input:checked ~ .label {
+  background: $color-lighter-grey;
 }
 
 .label--inline {
+  position: relative;
   font-weight: 400;
-  padding: 0 0.5rem;
-  width: calc(100% - 2.5rem);
-  display: inline-block;
-  vertical-align: text-top;
+  padding: 0.5rem 0.5rem 0.5rem 2rem;
+  display: block;
   margin: 0;
-  z-index: 1;
-  &::before {
-    transition: background-color 200ms ease-in-out;
-  }
+  cursor: pointer;
   @include fixed() {
     width: 25rem;
-  }
-}
-
-.input:focus,
-.input:checked,
-.input:hover {
-  + .label--inline {
-    &::before {
-      content: "";
-      position: absolute;
-      top: 0;
-      right: 0;
-      bottom: 0;
-      left: 0;
-      background: #f0f0f0;
-      width: calc(100% - 0.5rem);
-      height: 100%;
-      border-radius: 3px;
-      z-index: 0;
-    }
-  }
-}
-
-.input:checked {
-  + .label--inline {
-    &::before {
-      background: darken(#f0f0f0, 1%);
-    }
   }
 }
