+++ app/data/0_star_wars.json
@@ -43,7 +43,7 @@
                                         {
                                             "display": {
                                                 "properties": {
-                                                    "columns": true
+                                                    "columns": false
                                                 }
                                             },
                                             "guidance": "",
@@ -62,6 +62,13 @@
                                                 {
                                                     "label": "I prefer Star Trek",
                                                     "value": "I prefer Star Trek"
+                                                },
+                                                {
+                                                    "label": "Other",
+                                                    "value": "other",
+                                                    "other": {
+                                                      "label": "Please specify other"
+                                                    }
                                                 }
                                             ],
                                             "q_code": "20",
