+++ app/settings.py
@@ -50,7 +50,7 @@ EQ_LOG_LEVEL = os.getenv('EQ_LOG_LEVEL', 'INFO')
 EQ_CLOUDWATCH_LOGGING = parse_mode(os.getenv("EQ_CLOUDWATCH_LOGGING", 'True'))
 EQ_SCHEMA_DIRECTORY = os.getenv('EQ_SCHEMA_DIRECTORY', 'app/data')
 EQ_SESSION_TIMEOUT = int(os.getenv('EQ_SESSION_TIMEOUT', '28800'))
-EQ_SECRET_KEY = os.getenv('EQ_SECRET_KEY', os.urandom(24))
+EQ_SECRET_KEY = os.getenv('EQ_SECRET_KEY', 'test.test.test.test')
 EQ_UA_ID = os.getenv('EQ_UA_ID', '')
 EQ_SCHEMA_BUCKET = os.getenv('EQ_SCHEMA_BUCKET', "eq-schema-files-" + os.getenv('USER', 'UNKNOWN'))
 EQ_MAX_REPLAY_COUNT = os.getenv('EQ_REPLAY_COUNT', 50)
