+++ app/templates/partials/answer.html
@@ -11,7 +11,7 @@
 
   {%- set answer_fields %}
     {% if answer.schema_item.display and answer.schema_item.display.properties %}
-      {% set use_grid = answer.schema_item.display.properties.columns  %}
+      {% set use_grid = answer.schema_item.display.properties.columns %}
     {% endif %}
     {% include 'partials/answers/' ~ answer.schema_item.type|lower ~ '.html' %}
   {% endset -%}
