+++ app/templates/partials/answers/checkbox.html
@@ -1,3 +1,3 @@
-<div class="field field--multi field--{{answer.type|lower}} {{'field--cols' if use_grid}}">
+<div class="field field--checkbox {{'field--cols' if use_grid}}">
 {{ answer.schema_item.widget.render(answer) | safe}}
 </div>
