+++ app/templates/partials/forms/input.html
@@ -1,11 +1,10 @@
 {% set attr = {
-  "class": "input input--" ~ (input.valueType if input.valueType|length > 0 else input.type),
+  "class": "input",
   "type": input.type,
   "value": input.value,
   "name": input.name,
   "id": input.id,
-  "placeholder": input.placeholder,
-  "checked": "checked" if input.checked
+  "placeholder": input.placeholder
 } %}
 
 <input {{attr|xmlattr}} />
