+++ app/templates/partials/forms/label.html
@@ -5,5 +5,5 @@
 } %}
 
 <label {{attr|xmlattr}}>
-  <span class="label__inner">{{label.text}}</span>
+  {{label.text}}
 </label>
